backend/requirements.txt
livekit-agents>=1.3,<2
livekit-plugins-openai>=1.0,<2
livekit-plugins-cartesia>=1.3,<2
livekit-api>=1.1,<2
fastapi>=0.100
uvicorn>=0.20
pyyaml>=6.0
python-dotenv>=1.0
